+++ src/app/services/profesional.js
@@ -121,7 +121,7 @@ function profesionalServices() {
         if (avatarStatus.success) {
             if (avatarStatus.data) { 
                 let oldImage = destination + "/" + avatarStatus.data.foto;
-                if (fs.existsSync(oldImage)) {
+                if (fs.existsSync(oldImage) && avatarStatus.data.foto !== "defaultAvatar.png") {
                     fs.unlink(oldImage, (err) => {
                         if (err) throw err;
                         console.log(`deleted image: ${oldImage}`);
